Add popup for displaying off-screen matches
Replaces the status line with popup windows for displaying off-screen
matches. This employs the new, currently unstable, popup feature of Vim
version 8.1. The popup is displayed at the top of the window for matches
above the view, and respectively at the bottom for matches below. The
implementation is only a rough sketch without much thought to
configurability or code clarity.

One designated popup is created per-window and stored in a window-local
variable. It is then moved and showed/hid depending on the match or lack
thereof. Window movements are not handled.