Extract nvim-treesitter deprecations
nvim-treesitter is transitioning to being just a parsers manager,
with plugins being encouraged to move to independent initialization and
implementation.

This commit extracts nvim-treesitter plugin dependencies, copying those
too big to reimplement. Additionally, tree-sitter engine is reworked to
be an async library. When buffer's filetype is set, a parser is
initialized, a throttled reparser is attached to parser's on_bytes
event, and schedule_wrapped match extractor is attached to parser tree
change event. The latter extracts matches and stores them in a per-buffer
dictionary.

All requests to the treesitter backend now fetch data from the
aforementioned dictionary. Extra precautions are taken in get_matches
request, which now returns no matches if the buffer is known to have
changed with a parser that haven't caught up yet.